refactor(parser): collapse to a single Tokenizer, drop regex variant
Now that the state-machine tokenizer is the only one, remove the dual
implementation and present the codebase as if it were the first version.

- Delete `tokenizer-fsm.ts` and inline its content into `tokenizer.ts`,
  renaming the class from `TokenizerFSM` to `Tokenizer`.
- Remove `ITokenizer`, `TokenizerKind`, `ParserOptions`, `createTokenizer`
  factory from `parser.ts`. `Parser` constructor takes no options again
  and always uses `new Tokenizer(this)`.
- Delete `packages/parser/src/constants/specs.ts` (regex spec table) and
  the now-unused `Spec` type from `@vietscript/shared`.
- Drop `tokenizer-fsm.test.ts` (parity-vs-regex tests no longer apply);
  remaining tokenizer behavior is covered by tokenizer-edge,
  vietnamese-keywords, identifier-match-keyword, plus the snapshot
  drift / fixture parity smoke tests.
- Simplify benches: tokenizer.bench.ts now benches a single Tokenizer
  across the 5 fixtures; parser.bench.ts drops the "regex baseline" name.
- Drop stale comparison.json; baseline.json regenerated from the single
  tokenizer.

Docs:
- Rewrite `docs/architecture/tokenizer.md` to describe the current
  tokenizer as-is (no "switched from regex" framing, no comparison
  tables).
- README: replace the perf-comparison row with a neutral description of
  the tokenizer.
- roadmap.md / getting-started.md: drop comparison wording, link to the
  architecture page.
- CHANGELOG: collapse the migration entry into a neutral "Tokenizer"
  section describing the current design.

Tests: 357/357 pass. Lint + typecheck clean across 7 packages.
