+++ packages/parser/src/index.ts
@@ -6,9 +6,7 @@ export default parser
 
 export { VietScriptError } from './errors'
 export { Parser } from './parser'
-export type { ITokenizer, ParserOptions, TokenizerKind } from './parser'
 export { Tokenizer } from './tokenizer'
-export { TokenizerFSM } from './tokenizer-fsm'
 
 if (typeof window !== 'undefined') {
   (window as unknown as { VietScript: { parser: Parser } }).VietScript = { parser }
