+++ packages/shared/index.ts
@@ -1,5 +1,4 @@
 export * from './parser/keyword.enum'
 export * from './parser/node.interface'
 export * from './parser/operator.type'
-export * from './parser/spec.type'
 export * from './parser/token.type'
