--- packages/shared/parser/spec.type.ts
@@ -1,4 +1,0 @@
-import type { Keyword } from './keyword.enum'
-import type { Operator } from './operator.type'
-
-export type Spec = [RegExp, Keyword | Operator | null]
